+++ actix-server/CHANGES.md
@@ -5,7 +5,7 @@
 ## 2.6.0
 
 - Add `ServerBuilder::shutdown_signal()` method.
-- Minimum supported Rust version (MSRV) is now 1.80.
+- Minimum supported Rust version (MSRV) is now 1.75.
 
 ## 2.5.1