+++ actix-server/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.76.
+- Minimum supported Rust version (MSRV) is now 1.88.
 
 ## 2.6.0