+++ actix-server/CHANGES.md
@@ -4,6 +4,9 @@
 
 - Minimum supported Rust version (MSRV) is now 1.88.
 - internal: Use `core::future::{ready, Ready}` instead of actix-utils'
+- actix-server no longer removes the uds socket [#442]
+
+[#442]: https://github.com/actix/actix-net/pull/442
 
 ## 2.6.0